refactor(app): extract pure input translation and frame geometry
Issue #123 step 3. main.rs goes 3,152 -> 2,778 lines; the two extracted
modules take their tests with them, moved verbatim.

The issue is explicit that this is not permission to redesign behaviour, so
the check that matters is preservation rather than tidiness: the workspace
test count is identical before and after (936), 19 tests moved, and no
assertion was altered. Editing an assertion here would have meant the
extraction changed behaviour.

Only genuinely pure functions moved -- output depending solely on arguments.
Anything touching window state, the PTY, or the renderer's device objects
stayed behind for a later step; that is where the remaining concentration in
main.rs lives.

Both modules are declared in main.rs and their functions imported there, so
the extraction is reachable from the binary rather than a dead parallel copy.

Gates: fmt clean, clippy -D warnings clean, cargo test --workspace green.
