+++ crates/noren-app/src/main/tests.rs
@@ -4,18 +4,8 @@ use noren_app::passthrough::{self, collisions};
 use noren_app::session_persistence::{MAX_SESSION_STATE_BYTES, load};
 use noren_app::sidebar::EntryKind;
 
-#[test]
-fn winit_space_variants_encode_ascii_space() {
-    let variants = [
-        WinitKey::Named(NamedKey::Space),
-        WinitKey::Character(" ".into()),
-    ];
-    for logical_key in variants {
-        let input = translate_logical_key(&logical_key, KeyPhase::Pressed, Modifiers::empty())
-            .expect("space is supported terminal input");
-        assert_eq!(KeyEncoder::encode(input), Ok(vec![0x20]));
-    }
-}
+include!("../input_translation/tests.rs");
+include!("../frame_geometry/tests.rs");
 
 #[test]
 fn terminal_modes_drive_cursor_and_keypad_encoding() {
@@ -37,83 +27,6 @@ fn terminal_modes_drive_cursor_and_keypad_encoding() {
             .as_deref(),
         Ok(b"\x1bOq".as_slice())
     );
-}
-
-#[test]
-fn physical_keypad_mapping_is_bounded_to_numpad_codes() {
-    let cases = [
-        (KeyCode::Numpad0, KeypadKey::Zero),
-        (KeyCode::Numpad1, KeypadKey::One),
-        (KeyCode::Numpad2, KeypadKey::Two),
-        (KeyCode::Numpad3, KeypadKey::Three),
-        (KeyCode::Numpad4, KeypadKey::Four),
-        (KeyCode::Numpad5, KeypadKey::Five),
-        (KeyCode::Numpad6, KeypadKey::Six),
-        (KeyCode::Numpad7, KeypadKey::Seven),
-        (KeyCode::Numpad8, KeypadKey::Eight),
-        (KeyCode::Numpad9, KeypadKey::Nine),
-        (KeyCode::NumpadDecimal, KeypadKey::Decimal),
-        (KeyCode::NumpadAdd, KeypadKey::Plus),
-        (KeyCode::NumpadSubtract, KeypadKey::Minus),
-        (KeyCode::NumpadMultiply, KeypadKey::Star),
-        (KeyCode::NumpadDivide, KeypadKey::Slash),
-        (KeyCode::NumpadEnter, KeypadKey::Enter),
-    ];
-    for (code, expected) in cases {
-        assert_eq!(keypad_key(PhysicalKey::Code(code)), Some(expected));
-    }
-    assert_eq!(keypad_key(PhysicalKey::Code(KeyCode::Digit1)), None);
-}
-
-#[test]
-fn navigation_and_function_named_keys_translate_to_app_keys() {
-    let cases = [
-        (NamedKey::Delete, Key::Delete),
-        (NamedKey::Insert, Key::Insert),
-        (NamedKey::Home, Key::Home),
-        (NamedKey::End, Key::End),
-        (NamedKey::PageUp, Key::PageUp),
-        (NamedKey::PageDown, Key::PageDown),
-        (NamedKey::F1, Key::Function(FunctionKey::F1)),
-        (NamedKey::F2, Key::Function(FunctionKey::F2)),
-        (NamedKey::F3, Key::Function(FunctionKey::F3)),
-        (NamedKey::F4, Key::Function(FunctionKey::F4)),
-        (NamedKey::F5, Key::Function(FunctionKey::F5)),
-        (NamedKey::F6, Key::Function(FunctionKey::F6)),
-        (NamedKey::F7, Key::Function(FunctionKey::F7)),
-        (NamedKey::F8, Key::Function(FunctionKey::F8)),
-        (NamedKey::F9, Key::Function(FunctionKey::F9)),
-        (NamedKey::F10, Key::Function(FunctionKey::F10)),
-        (NamedKey::F11, Key::Function(FunctionKey::F11)),
-        (NamedKey::F12, Key::Function(FunctionKey::F12)),
-    ];
-    for (named, expected) in cases {
-        let logical_key = WinitKey::Named(named);
-        let input = translate_logical_key(&logical_key, KeyPhase::Pressed, Modifiers::empty())
-            .expect("stage one key is supported terminal input");
-        assert_eq!(input.key(), expected);
-        assert_eq!(input.phase(), KeyPhase::Pressed);
-    }
-}
-
-#[test]
-fn untranslated_named_keys_still_report_a_drop() {
-    for named in [NamedKey::F13, NamedKey::ScrollLock, NamedKey::Pause] {
-        let logical_key = WinitKey::Named(named);
-        assert_eq!(
-            translate_logical_key(&logical_key, KeyPhase::Pressed, Modifiers::empty()),
-            Err(KeyDropReason::UnsupportedKey)
-        );
-    }
-}
-
-#[test]
-fn pixel_row_index_truncates_and_rejects_non_finite() {
-    assert_eq!(pixel_row_index(0.0, 20), Some(0));
-    assert_eq!(pixel_row_index(39.0, 20), Some(1));
-    assert_eq!(pixel_row_index(40.0, 20), Some(2));
-    assert_eq!(pixel_row_index(f64::NAN, 20), None);
-    assert_eq!(pixel_row_index(f64::INFINITY, 20), None);
 }
 
 #[test]
@@ -394,47 +307,6 @@ fn terminal_event_finishes_the_session_without_closing_the_window() {
 }
 
 #[test]
-fn diagnostics_chord_is_a_super_d_press_only() {
-    let super_modifiers = Modifiers::empty().super_key();
-    let chord = WinitKey::Character("d".into());
-    for (state, repeat, modifiers, expected) in [
-        (ElementState::Pressed, false, super_modifiers, true),
-        (ElementState::Released, false, super_modifiers, false),
-        (ElementState::Pressed, true, super_modifiers, false),
-        (ElementState::Pressed, false, Modifiers::empty(), false),
-        (
-            ElementState::Pressed,
-            false,
-            Modifiers::empty().shift(),
-            false,
-        ),
-    ] {
-        assert_eq!(
-            diagnostics_chord_pressed(&chord, state, repeat, modifiers),
-            expected,
-            "state={state:?} repeat={repeat}"
-        );
-    }
-    for other in [
-        WinitKey::Character("x".into()),
-        WinitKey::Character("dd".into()),
-        WinitKey::Named(NamedKey::Enter),
-    ] {
-        assert!(
-            !diagnostics_chord_pressed(&other, ElementState::Pressed, false, super_modifiers),
-            "only D toggles diagnostics"
-        );
-    }
-    let shifted = WinitKey::Character("D".into());
-    assert!(diagnostics_chord_pressed(
-        &shifted,
-        ElementState::Pressed,
-        false,
-        super_modifiers
-    ));
-}
-
-#[test]
 fn toggle_diagnostics_reports_live_state_and_clears_on_exit() {
     let mut app = NorenApp::default();
     let mut terminal = TerminalState::new(4, 8).expect("valid terminal");
@@ -659,210 +531,6 @@ fn palette_carries_all_four_canonical_commands() {
     );
 }
 
-// =========================================================================
-// Sidebar geometry: the terminal width, the PTY winsize, and the renderer's
-// drawn region must all agree once the sidebar reserves 16 columns.
-// =========================================================================
-
-/// Number of terminal cell columns the renderer drew, measured from its
-/// vertex output rather than restating the column formula. Each terminal
-/// column is fed a glyph the renderer lights starting at the cell's left
-/// pixel edge (`B` lights glyph column 0), so a drawn column is detectable
-/// as a glyph rect whose LEFT edge sits on that boundary. Scanning runs
-/// rightward from the first terminal column (`SIDEBAR_COLS`) until a column
-/// has no glyph — terminal content is contiguous, so the first gap marks
-/// the end of the drawn region.
-///
-/// Matching a rect's left edge (its top-left corner) — not *any* vertex on
-/// the boundary — is essential: a glyph's rightmost lit pixel column (e.g.
-/// `B`, whose rows `17 = 0b10001` light glyph column 4) produces a rect
-/// whose RIGHT edge lands exactly on the next column's left edge. Matching
-/// arbitrary vertices would count that bleed as a drawn column and over-
-/// count by one. Each rect is emitted as a 6-vertex fan whose first vertex
-/// is its top-left corner, so the left edges are read from every 6th group.
-fn rendered_terminal_columns(vertices: &[renderer::Vertex], width: u32, cell_width: u32) -> usize {
-    let rect_lefts: Vec<f32> = vertices
-        .chunks_exact(6)
-        .map(|rect| rect[0].position[0])
-        .collect();
-    let mut drawn = 0;
-    for col in renderer::SIDEBAR_COLS..usize::from(MAX_RENDER_COLS) {
-        let edge = ((col as u32) * cell_width) as f32 / width as f32 * 2.0 - 1.0;
-        if rect_lefts.iter().any(|left| (left - edge).abs() < 1e-5) {
-            drawn += 1;
-        } else {
-            break;
-        }
-    }
-    drawn
-}
-
-/// Drive the three terminal-width consumers — `TerminalState`'s stored
-/// column count, the PTY winsize, and the columns the renderer actually
-/// draws — at one window width and cell size, asserting they all agree on
-/// `terminal_cols(window_cols)`. Asserting `terminal_cols() == window_cols -
-/// 16` would merely restate the formula and pass any consistent-but-wrong
-/// value; instead this exercises the three real consumers and is shared by
-/// the swept agreement test below across every regime where they can drift
-/// apart — including non-default cell sizes.
-fn assert_three_consumers_agree_at(width: u32, metrics: CellMetrics) {
-    let height = 600_u32;
-    let cell_width = metrics.width();
-    let cell_height = metrics.height();
-    let window_cols = u16::try_from(width / cell_width).expect("fits in u16");
-    let cols = terminal_cols(window_cols);
-
-    // Consumer 1: the terminal state stores the sidebar-adjusted width.
-    let rows = u16::try_from(height / cell_height).expect("fits in u16");
-    let mut terminal = TerminalState::new(rows, cols).expect("valid terminal");
-    terminal.feed_bytes(&vec![b'B'; usize::from(cols)]);
-    let (_, term_cols) = terminal.size();
-    assert_eq!(
-        term_cols,
-        cols,
-        "at {width}px cell {}x{}: terminal must store \
-             terminal_cols({window_cols}) = {cols}",
-        metrics.width(),
-        metrics.height(),
-    );
-
-    // Consumer 2: the PTY winsize carries the same column count.
-    let pty = PtySize::from_raw(rows, cols).expect("valid pty size");
-    assert_eq!(
-        pty.cols(),
-        cols,
-        "at {width}px cell {}x{}: PTY winsize must agree",
-        metrics.width(),
-        metrics.height(),
-    );
-
-    // Consumer 3: the renderer draws exactly that many terminal columns —
-    // measured from vertex output, independent of `terminal_cols`. The cell
-    // metrics are threaded through so a renderer still drawing at the
-    // compile-time default is exposed at a non-default size.
-    let snapshot = terminal.snapshot();
-    let sidebar: Vec<String> = Vec::new();
-    let vertices = renderer::glyph_vertices(
-        Some(&snapshot),
-        Some(sidebar.as_slice()),
-        None,
-        width,
-        height,
-        metrics,
-    );
-    let drawn = rendered_terminal_columns(&vertices, width, cell_width);
-    assert_eq!(
-        drawn,
-        usize::from(cols),
-        "at {width}px cell {}x{} (window_cols={window_cols}): \
-             renderer drew {drawn} terminal columns but terminal/PTY agree on {cols} — \
-             the sidebar width is not consistently subtracted, the upper clamp is missing, \
-             or the renderer ignored the configured cell size",
-        metrics.width(),
-        metrics.height(),
-    );
-}
-
-/// The PR's headline property: the three consumers agree once the sidebar
-/// reserves 16 columns — swept across the input range rather than pinned at
-/// one width. A single point cannot support "agreement across the range",
-/// and the original 900px point sits squarely inside the band (17..=160
-/// columns) where the pre-fix geometry already agreed. Each swept width
-/// targets a distinct regime so a regression of either pre-fix defect is
-/// caught:
-///   - 80px   (8 cols, below the sidebar width): the floor regime. A
-///     renderer that floored at zero terminal columns while the terminal
-///     and PTY held one is exposed.
-///   - 900px  (90 cols, a typical window): the common case — the very width
-///     a prior commit message wrongly cited as the divergence (both the
-///     pre- and post-fix geometry agree at 74 here).
-///   - 1600px (160 cols, exactly `MAX_RENDER_COLS`): the budget boundary,
-///     where the terminal fills the whole drawable budget (144).
-///   - 2000px (200 cols, above `MAX_RENDER_COLS`): the upper-clamp regime.
-///     A `terminal_cols` with no upper clamp would claim 184 columns while
-///     the renderer draws 144, silently clipping 40 columns of output.
-///
-/// This is a regression guard, not a reproduced bug: at the moment this
-/// test was added all three consumers already agreed at every swept width,
-/// and it exists to hold that line. Mutating `terminal_cols` to drop the
-/// sidebar subtraction breaks the typical/above-max widths, and dropping
-/// the upper clamp breaks the 2000px width.
-///
-/// The test is also swept across **cell sizes**. Issue #76: a configured
-/// `cell_width = 20` flows to the geometry/PTY but the renderer ignored it,
-/// drawing at the 10px compile-time constant. At 20px every width in the
-/// sweep produces half the window_cols, so the renderer — if still on the
-/// constant — would draw *twice* as many terminal columns as the terminal
-/// and PTY agree on, and the three consumers diverge. This is the
-/// acceptance criterion from the Issue.
-#[test]
-fn terminal_cols_pty_winsize_and_renderer_agree_across_the_width_range() {
-    let poc = GridGeometry::poc().cell_metrics();
-    for width in [80_u32, 900, 1600, 2000] {
-        assert_three_consumers_agree_at(width, poc);
-    }
-    let big = GridGeometry::with_cells(20, 40)
-        .expect("valid geometry")
-        .cell_metrics();
-    for width in [160_u32, 1800, 3200, 4000] {
-        assert_three_consumers_agree_at(width, big);
-    }
-}
-
-#[test]
-fn terminal_rows_pty_winsize_and_renderer_agree_with_permanent_status_chrome() {
-    for window_rows in [1_u16, 2, 30, noren_app::MAX_RENDER_ROWS] {
-        let mut app = NorenApp {
-            status: "Noren PoC ready",
-            show_status: false,
-            ..Default::default()
-        };
-        let metrics = app.geometry.cell_metrics();
-        let height = u32::from(window_rows) * metrics.height();
-        let grid = app
-            .geometry
-            .update(Resize::new(WINDOW_WIDTH, height))
-            .expect("non-zero window grid");
-        assert_eq!(grid.rows(), window_rows);
-
-        // Drive the exact initialization seam used before PtySession::spawn
-        // instead of rebuilding its dimensions inside the test.
-        let pty = app
-            .prepare_initial_terminal(grid)
-            .expect("valid runtime grid");
-        let terminal = app.terminal.as_ref().expect("terminal installed");
-        let terminal_rows = terminal.size().0;
-        let status = app.rendered_status_row(window_rows);
-        let layout = renderer::FrameRowLayout::new(
-            height,
-            metrics,
-            usize::from(terminal_rows),
-            status.is_some(),
-        )
-        .expect("non-zero frame layout");
-
-        assert_eq!(terminal.size().0, terminal_rows);
-        assert_eq!(pty.rows(), terminal_rows);
-        assert_eq!(layout.row_at(0), Some(renderer::FrameRow::Terminal(0)));
-        assert_eq!(
-            layout.row_at(usize::from(terminal_rows - 1)),
-            Some(renderer::FrameRow::Terminal(usize::from(terminal_rows - 1)))
-        );
-        if window_rows == 1 {
-            assert_eq!(status, None);
-            assert_eq!(layout.rendered_rows(), 1);
-        } else {
-            assert!(status.is_some());
-            assert_eq!(terminal_rows, window_rows - 1);
-            assert_eq!(
-                layout.row_at(usize::from(window_rows - 1)),
-                Some(renderer::FrameRow::Status)
-            );
-            assert_eq!(layout.rendered_rows(), usize::from(window_rows));
-        }
-    }
-}
-
 #[test]
 fn pending_resize_applies_the_runtime_status_row_contract_to_terminal_state() {
     let mut app = NorenApp {
@@ -883,44 +551,6 @@ fn pending_resize_applies_the_runtime_status_row_contract_to_terminal_state() {
     assert_eq!(
         app.terminal.as_ref().expect("terminal retained").size(),
         (29, terminal_cols(grid.cols()))
-    );
-}
-
-/// MINOR-1: below ~160px the window fits inside the sidebar. `terminal_cols`
-/// floors at one (the terminal/PTY reject zero columns); the renderer must
-/// floor at the same one rather than drawing zero terminal columns while the
-/// terminal still holds one. Drives the real renderer so the agreement is
-/// measured, not assumed.
-#[test]
-fn terminal_cols_and_renderer_floor_at_one_below_the_sidebar() {
-    let cell_width = GridGeometry::poc().cell_width();
-    // A window exactly SIDEBAR_COLS wide: visible_cols == SIDEBAR_COLS, so
-    // the terminal region has no room — both floors must keep it at one.
-    let width = (renderer::SIDEBAR_COLS as u32) * cell_width;
-    let height = 600_u32;
-    let window_cols = u16::try_from(width / cell_width).expect("fits in u16");
-    assert_eq!(window_cols, u16::try_from(renderer::SIDEBAR_COLS).unwrap());
-    let cols = terminal_cols(window_cols);
-    assert_eq!(cols, 1, "terminal_cols floors at one, never zero");
-
-    let mut terminal = TerminalState::new(2, cols).expect("valid terminal");
-    terminal.feed_bytes(&vec![b'B'; usize::from(cols)]);
-    let snapshot = terminal.snapshot();
-    let sidebar: Vec<String> = Vec::new();
-    let vertices = renderer::glyph_vertices(
-        Some(&snapshot),
-        Some(sidebar.as_slice()),
-        None,
-        width,
-        height,
-        GridGeometry::poc().cell_metrics(),
-    );
-    let drawn = rendered_terminal_columns(&vertices, width, cell_width);
-    assert_eq!(
-        drawn,
-        usize::from(cols),
-        "renderer must draw the terminal's one column, not zero — the floor \
-             disagrees with terminal_cols below the sidebar width"
     );
 }
 
@@ -977,112 +607,6 @@ fn sidebar_text_lines_format_a_real_workspace_sidebar() {
     );
 }
 
-/// MINOR-3: `grid_point_at`'s sidebar boundary. A click in the last sidebar
-/// column must be rejected, and a click at the first terminal column must
-/// map to terminal cell 0. `grid_point_at` itself needs a live window this
-/// harness cannot create, so this drives the extracted column mapper that
-/// `grid_point_at` delegates to.
-#[test]
-fn terminal_column_at_rejects_the_sidebar_and_starts_the_terminal_at_zero() {
-    let cols = 40_u16;
-    let cell_width = GridGeometry::poc().cell_width();
-    let sidebar_edge = sidebar_pixel_width(cell_width);
-
-    // The last sidebar column — just inside the sidebar's right edge — does
-    // not address the terminal grid.
-    assert_eq!(
-        terminal_column_at(sidebar_edge - 1.0, cols, cell_width),
-        None,
-        "a click in the last sidebar column must be rejected"
-    );
-    // The first terminal column, exactly at the sidebar's right edge, maps
-    // to terminal cell 0.
-    assert_eq!(
-        terminal_column_at(sidebar_edge, cols, cell_width),
-        Some(0),
-        "the first terminal column must map to cell 0"
-    );
-    // One cell width further in lands in terminal cell 1.
-    assert_eq!(
-        terminal_column_at(sidebar_edge + f64::from(cell_width), cols, cell_width),
-        Some(1)
-    );
-    // The last terminal column maps to the highest valid cell.
-    assert_eq!(
-        terminal_column_at(
-            sidebar_edge + f64::from(cell_width) * f64::from(cols - 1),
-            cols,
-            cell_width
-        ),
-        Some(usize::from(cols - 1))
-    );
-    // A click past the last column clamps to the last cell, never overflows.
-    assert_eq!(
-        terminal_column_at(
-            sidebar_edge + f64::from(cell_width) * f64::from(cols),
-            cols,
-            cell_width
-        ),
-        Some(usize::from(cols - 1))
-    );
-    // Negative and non-finite clicks are rejected.
-    assert_eq!(terminal_column_at(-1.0, cols, cell_width), None);
-    assert_eq!(terminal_column_at(f64::NAN, cols, cell_width), None);
-}
-
-/// Issue #76: at a non-default cell width, the sidebar's drawn pixel
-/// boundary and the click-handling boundary must agree. The renderer draws
-/// the terminal starting at column `SIDEBAR_COLS`; `sidebar_pixel_width`
-/// and `terminal_column_at` must use the same `cell_width` to locate that
-/// boundary, or clicks land in the wrong region.
-///
-/// At `cell_width = 20` the boundary is `16 * 20 = 320px`. If
-/// `sidebar_pixel_width` still used `POC_CELL_WIDTH` (10), the boundary
-/// would drift to 160px and clicks in the 160–320px strip would be
-/// misattributed to the terminal instead of the sidebar.
-#[test]
-fn sidebar_boundary_and_click_boundary_agree_at_non_default_cell_width() {
-    let cell_width = 20_u32;
-    let cols = 40_u16;
-    let sidebar_edge = sidebar_pixel_width(cell_width);
-
-    // The drawn boundary is SIDEBAR_COLS * cell_width.
-    assert_eq!(
-        sidebar_edge,
-        f64::from((renderer::SIDEBAR_COLS as u32) * cell_width),
-        "sidebar pixel width must be SIDEBAR_COLS * cell_width"
-    );
-    assert_eq!(sidebar_edge, 320.0);
-
-    // A click at the boundary maps to terminal column 0.
-    assert_eq!(
-        terminal_column_at(sidebar_edge, cols, cell_width),
-        Some(0),
-        "at cell_width=20, the first terminal column is at the sidebar edge"
-    );
-    // A click one pixel left of the boundary is still the sidebar.
-    assert_eq!(
-        terminal_column_at(sidebar_edge - 1.0, cols, cell_width),
-        None,
-        "at cell_width=20, a click just left of the boundary is the sidebar"
-    );
-    // One cell width further maps to column 1.
-    assert_eq!(
-        terminal_column_at(sidebar_edge + f64::from(cell_width), cols, cell_width),
-        Some(1),
-        "at cell_width=20, one cell past the boundary is column 1"
-    );
-    // The last terminal column maps to the highest valid cell.
-    assert_eq!(
-        terminal_column_at(
-            sidebar_edge + f64::from(cell_width) * f64::from(cols - 1),
-            cols,
-            cell_width
-        ),
-        Some(usize::from(cols - 1)),
-    );
-}
-
 // ── Pass-through gate integration tests ──────────────────────────────
 
 /// The palette policy claims exactly two chords: Super+Escape (exit) and
@@ -1105,104 +629,6 @@ fn palette_policy_claims_exactly_super_escape_and_super_p() {
         super_p,
         "palette claim is Super+p"
     );
-}
-
-/// A multi-chord leader whose first chord is held and then mismatched must
-/// replay the held chord byte-for-byte before the mismatching chord. This
-/// is the replay path whose failure would silently break Zellij.
-#[test]
-fn leader_mismatch_replays_held_chord_bytes_in_order() {
-    // A two-chord palette leader on chords absent from the Zellij corpus:
-    // bare 'a' then 'g'. Both encode non-empty bytes so a lost or
-    // reordered replay changes what the child receives.
-    let claim = PassthroughClaim {
-        id: CLAIM_ID_PALETTE,
-        action: PassthroughAction::OpenCommandPalette,
-        seq: ChordSeq::new(vec![
-            Chord::new(GateKeyCode::Char('a'), GateModifiers::empty()).unwrap(),
-            Chord::new(GateKeyCode::Char('g'), GateModifiers::empty()).unwrap(),
-        ])
-        .unwrap(),
-        justification: "test",
-    };
-    let policy = PassthroughPolicy::try_new(vec![
-        PassthroughClaim {
-            id: passthrough::CLAIM_ID_EXIT,
-            action: PassthroughAction::ExitToWorkspace,
-            seq: ChordSeq::single(
-                Chord::new(GateKeyCode::Escape, GateModifiers::empty().super_key()).unwrap(),
-            ),
-            justification: "test",
-        },
-        claim,
-    ])
-    .expect("valid manifest");
-
-    let mode = InputMode::normal();
-    let mut gate = PassthroughGate::new();
-
-    // Press 'a': the first chord of the palette leader — held as pending.
-    let chord_a = Chord::new(GateKeyCode::Char('a'), GateModifiers::empty()).expect("normalized");
-    let decision = gate.press(&policy, chord_a);
-    assert_eq!(decision.kind, GateKind::Pending);
-    assert!(decision.replayed.is_empty(), "pending must not replay");
-
-    // Press 'x': not the second chord. The gate forwards, replaying 'a'
-    // first. The replay must arrive byte-for-byte before 'x'.
-    let chord_x = Chord::new(GateKeyCode::Char('x'), GateModifiers::empty()).expect("normalized");
-    let decision = gate.press(&policy, chord_x);
-    assert_eq!(decision.kind, GateKind::Forwarded);
-    assert_eq!(
-        decision.replayed,
-        vec![chord_a],
-        "the held leader chord must be replayed"
-    );
-
-    // Verify the replay bytes match direct encoding.
-    let replay_bytes = encode_chord(&decision.replayed[0], mode).expect("encodes");
-    assert_eq!(replay_bytes, b"a", "replayed 'a' must encode to b\"a\"");
-
-    // After the mismatch, the gate is clean: a subsequent 'x' forwards
-    // with no replay.
-    let decision = gate.press(
-        &policy,
-        Chord::new(GateKeyCode::Char('x'), GateModifiers::empty()).unwrap(),
-    );
-    assert_eq!(decision.kind, GateKind::Forwarded);
-    assert!(decision.replayed.is_empty(), "gate is clean after mismatch");
-}
-
-/// When the palette is closed, the gate forwards unclaimed chords. The
-/// forwarded encoding must match what the encoder produces directly —
-/// byte-identical to the pre-gate behaviour.
-#[test]
-fn closed_palette_forwarded_key_is_byte_identical_to_direct_encode() {
-    let policy = palette_policy(KeymapConfig::default());
-    let mode = InputMode::normal();
-    let mut gate = PassthroughGate::new();
-
-    for (code, modifiers) in [
-        (GateKeyCode::Char('a'), GateModifiers::empty()),
-        (GateKeyCode::Char('z'), GateModifiers::empty()),
-        (GateKeyCode::Enter, GateModifiers::empty()),
-        (GateKeyCode::Char('c'), GateModifiers::empty().ctrl()),
-        (GateKeyCode::Char('f'), GateModifiers::empty().alt()),
-    ] {
-        let chord = Chord::new(code, modifiers).expect("normalized");
-        let decision = gate.press(&policy, chord);
-        assert_eq!(decision.kind, GateKind::Forwarded, "chord must forward");
-        let forwarded = encode_chord(&chord, mode).expect("encodes");
-
-        let app_key = gate_key_to_app(code).expect("maps to app key");
-        let app_mods = app_modifiers_from_gate(modifiers);
-        let direct =
-            KeyEncoder::encode_with(KeyInput::new(app_key, KeyPhase::Pressed, app_mods), mode)
-                .expect("encodes");
-        assert_eq!(
-            forwarded, direct,
-            "forwarded bytes must match direct encode for {code:?}"
-        );
-    }
 }
 
 /// Super+p is intercepted by the gate (opens the palette) and produces no
@@ -1837,56 +1263,6 @@ fn shift_bypasses_tracking_for_local_selection() {
     assert!(!app.mouse_reportable(), "Shift bypasses tracking");
 }
 
-// ── Sidebar offset and coordinate mapping ───────────────────────────
-
-/// A click on the first terminal column (exactly at the sidebar edge)
-/// reports column 1, not 17. This exercises the sidebar subtraction in
-/// `terminal_column_at` through the encoder's 1-based conversion — if the
-/// sidebar offset were dropped, the column would be 17 (16 sidebar cells
-/// + 1).
-#[test]
-fn sidebar_offset_first_terminal_column_reports_col_1() {
-    let cols = 40_u16;
-    let cell_width = GridGeometry::poc().cell_width();
-    let col = terminal_column_at(sidebar_pixel_width(cell_width), cols, cell_width)
-        .expect("first terminal column must map to a cell");
-    assert_eq!(col, 0, "sidebar offset: first terminal column = cell 0");
-
-    let grid = MouseGrid::new(10, cols).expect("grid");
-    let modes = MouseModes::disabled().with_normal(true).with_sgr(true);
-    let event = PointerEvent::press(
-        EncoderButton::Left,
-        col as u32,
-        0,
-        PointerModifiers::empty(),
-    );
-    let bytes = MouseEncoder::encode(event, modes, grid).expect("must encode");
-    let report = String::from_utf8(bytes).expect("SGR is ASCII");
-    assert_eq!(
-        report, "\x1b[<0;1;1M",
-        "column must be 1 (sidebar offset applied), not 17"
-    );
-}
-
-/// A click inside the sidebar produces no terminal column, hence no PTY
-/// bytes can be constructed for it.
-#[test]
-fn sidebar_click_produces_no_terminal_column() {
-    let cols = 40_u16;
-    let cell_width = GridGeometry::poc().cell_width();
-    let edge = sidebar_pixel_width(cell_width);
-    assert_eq!(
-        terminal_column_at(edge - 1.0, cols, cell_width),
-        None,
-        "last sidebar column must not map to a terminal cell"
-    );
-    assert_eq!(
-        terminal_column_at(0.0, cols, cell_width),
-        None,
-        "leftmost pixel is sidebar"
-    );
-}
-
 // ── Encoder integration: tracking modes ─────────────────────────────
 
 /// Mode 1000 with X10 byte form: a left click at (0,0) produces
@@ -1975,141 +1351,6 @@ fn no_modes_means_no_bytes_for_any_mouse_event() {
             "{label} must produce nothing with no modes"
         );
     }
-}
-
-// ── Button and wheel mapping ────────────────────────────────────────
-
-/// Left/Middle/Right map to the encoder's button enum; Back/Forward/Other
-/// return None and are never reported.
-#[test]
-fn encode_button_maps_known_and_ignores_extended() {
-    assert_eq!(encode_button(MouseButton::Left), Some(EncoderButton::Left));
-    assert_eq!(
-        encode_button(MouseButton::Middle),
-        Some(EncoderButton::Middle)
-    );
-    assert_eq!(
-        encode_button(MouseButton::Right),
-        Some(EncoderButton::Right)
-    );
-    assert_eq!(encode_button(MouseButton::Back), None);
-    assert_eq!(encode_button(MouseButton::Forward), None);
-    assert_eq!(encode_button(MouseButton::Other(1)), None);
-}
-
-/// Wheel delta sign maps to direction, and magnitude maps to click count.
-///
-/// winit 0.30 `MouseScrollDelta` docs (from the source at
-/// `winit-0.30.13/src/event.rs`):
-///
-///   LineDelta:  "Positive values indicate that the content that is being
-///                scrolled should move right and down (revealing more
-///                content left and up)."
-///   PixelDelta: "Positive values indicate that the content being scrolled
-///                should move right/down."
-///
-/// Positive y = content moves down = the user scrolled **up** (xterm button
-/// 4, `Cb=64`). Negative y = scroll down (`Cb=65`). Both variants share the
-/// same sign convention.
-#[test]
-fn wheel_clicks_direction_and_count() {
-    let metrics = GridGeometry::poc().cell_metrics();
-    // LineDelta: positive y = wheel up (content moves down, revealing
-    // earlier content). See the winit sentence quoted above.
-    let up = wheel_clicks(MouseScrollDelta::LineDelta(0.0, 1.0), metrics);
-    assert_eq!(up, vec![WheelDirection::Up]);
-
-    // Negative y = wheel down.
-    let down = wheel_clicks(MouseScrollDelta::LineDelta(0.0, -3.0), metrics);
-    assert_eq!(down, vec![WheelDirection::Down; 3]);
-
-    // PixelDelta shares the same sign convention: positive y = wheel up.
-    let pixel_up = wheel_clicks(
-        MouseScrollDelta::PixelDelta(PhysicalPosition::new(
-            0.0,
-            f64::from(metrics.height()) * 2.0,
-        )),
-        metrics,
-    );
-    assert_eq!(pixel_up, vec![WheelDirection::Up; 2]);
-
-    // Pin the full path through the encoder so the emitted bytes are fixed:
-    // wheel up → Cb=64, wheel down → Cb=65 (SGR form, 1-based col/row).
-    let grid = MouseGrid::new(10, 40).expect("grid");
-    let modes = MouseModes::disabled().with_normal(true).with_sgr(true);
-    let up_event = PointerEvent::wheel(up[0], 0, 0, PointerModifiers::empty());
-    assert_eq!(
-        MouseEncoder::encode(up_event, modes, grid).as_deref(),
-        Some(b"\x1b[<64;1;1M".as_slice()),
-        "wheel up must emit Cb=64"
-    );
-    let down_event = PointerEvent::wheel(down[0], 0, 0, PointerModifiers::empty());
-    assert_eq!(
-        MouseEncoder::encode(down_event, modes, grid).as_deref(),
-        Some(b"\x1b[<65;1;1M".as_slice()),
-        "wheel down must emit Cb=65"
-    );
-}
-
-/// A resting trackpad reports `PixelDelta{y:0}`, which is a genuine zero
-/// delta and must produce no wheel reports. The `handle_mouse_wheel` loop
-/// consumes the returned vec, so an empty vec means no bytes are written to
-/// the PTY — a spurious `Down` click would corrupt the application.
-#[test]
-fn wheel_clicks_zero_delta_produces_nothing() {
-    let metrics = GridGeometry::poc().cell_metrics();
-    // PixelDelta zero — the resting-trackpad case the bug shipped on.
-    let pixel_zero = wheel_clicks(
-        MouseScrollDelta::PixelDelta(PhysicalPosition::new(0.0, 0.0)),
-        metrics,
-    );
-    assert!(pixel_zero.is_empty(), "zero PixelDelta must emit nothing");
-
-    // LineDelta zero must likewise emit nothing.
-    let line_zero = wheel_clicks(MouseScrollDelta::LineDelta(0.0, 0.0), metrics);
-    assert!(line_zero.is_empty(), "zero LineDelta must emit nothing");
-}
-
-/// A `PixelDelta` must convert pixels to lines at the **configured** cell
-/// height, never a compile-time constant. Built the same way the app builds
-/// its geometry — from parsed configuration — at `cell_height = 40`, double
-/// the PoC default of 20. A 40px scroll is exactly one line here; if
-/// `wheel_clicks` divided by the hardcoded default it would yield two. This
-/// is the guard that stops the constant creeping back into the pixel→line
-/// conversion.
-#[test]
-fn wheel_clicks_pixel_delta_uses_configured_cell_height() {
-    let config = AppConfig::parse("[font]\ncell_height = 40\n").expect("valid configuration");
-    let metrics = GridGeometry::with_cells(config.font().cell_width(), config.font().cell_height())
-        .expect("valid geometry")
-        .cell_metrics();
-    assert_eq!(metrics.height(), 40);
-
-    // One configured cell height of pixels up = exactly one wheel-up click.
-    // A hardcoded POC_CELL_HEIGHT (20) would divide 40px into two clicks.
-    let one_line_up = wheel_clicks(
-        MouseScrollDelta::PixelDelta(PhysicalPosition::new(0.0, f64::from(metrics.height()))),
-        metrics,
-    );
-    assert_eq!(
-        one_line_up,
-        vec![WheelDirection::Up],
-        "at cell_height=40, 40px is one line, not the two a hardcoded 20 would give"
-    );
-
-    // Two configured cell heights down = exactly two wheel-down clicks.
-    let two_lines_down = wheel_clicks(
-        MouseScrollDelta::PixelDelta(PhysicalPosition::new(
-            0.0,
-            -f64::from(metrics.height()) * 2.0,
-        )),
-        metrics,
-    );
-    assert_eq!(
-        two_lines_down,
-        vec![WheelDirection::Down; 2],
-        "at cell_height=40, -80px is two lines, not the four a hardcoded 20 would give"
-    );
 }
 
 // ── Selection preservation with tracking disabled ───────────────────
